64-bit fixes

+++ src/libreplace/strncasecmp.c
@@ -8,7 +8,7 @@
 #include <ctype.h>
 
 int
-strncasecmp(const char *s1, const char *s2, unsigned int n)
+strncasecmp(const char *s1, const char *s2, size_t n)
 {
 	if (n == 0)
 		return 0;
